chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "devyoon",email = "wideskyinme@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.15,<0.4.0)",
    "langchain-core (>=0.3.31,<0.4.0)",
    "langchain-chroma (>=0.2.0,<0.3.0)",
    "langchain-community (>=0.3.15,<0.4.0)",
    "langchain-openai (>=0.3.1,<0.4.0)",
    "langchain-pinecone (>=0.2.2,<0.3.0)",
    "langchain-text-splitters (>=0.3.5,<0.4.0)",
    "langchain-upstage (>=0.6.0,<0.7.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langsmith (>=0.3.0,<0.4.0)",
    "langchain-upstage (>=0.6.0,<0.7.0)",
    "langchain-pinecone (>=0.2.2,<0.3.0)",
    "protobuf (>=5.0,<6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
